% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDiscreteTime.R
\name{ddtfm}
\alias{ddtfm}
\title{Probability density function of fitted model}
\usage{
ddtfm(r, distr = "norm", param)
}
\arguments{
\item{r}{region of returns}

\item{distr}{distribution name: "norm" or "unif"}

\item{param}{parameters of distribution as a vector}
}
\value{
numeric
}
\description{
Probability density function of fitted model
}
