% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-family.R
\name{dgbm}
\alias{dgbm}
\title{#' Probability density function for Black-Scholes distribution}
\usage{
dgbm(s, t, spot, rate, volat)
}
\arguments{
\item{s}{real-numeric input to PDF representing price level}

\item{t}{positive time value}

\item{spot}{the initial spot price}

\item{rate}{the risk-free rate of return}

\item{volat}{the volatility level}
}
\value{
numeric
}
\description{
{The probability density function for Black-Scholes distributed
variables.}
}
