% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricCompensatedPoissonProcess.R
\name{dgcpp}
\alias{dgcpp}
\title{Cumulative distribution function for geometric compensated Poisson processes}
\usage{
dgcpp(s, t, spot, a, b, lambda)
}
\arguments{
\item{s}{price argument to CDF}

\item{t}{time argument to CDF}

\item{spot}{initial spot price}

\item{a}{the coefficient of the Poisson RV}

\item{b}{the coefficient of the compensated drift}

\item{lambda}{the mean rate of arrivals}
}
\value{
numeric
}
\description{
{The CDF of a geometrically compensated Poisson process at a given
time and initial spot price.}
}
