% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbmHittingTimes.R
\name{dhit_gbm}
\alias{dhit_gbm}
\title{The PDF of the hitting time of for a GBM to reach a price level}
\usage{
dhit_gbm(t, gbm, target, spot)
}
\arguments{
\item{t}{time input}

\item{gbm}{the parameters defining the GBM, a drift and volatility}

\item{target}{the target price to reach in \code{t} time}

\item{spot}{the spot price}
}
\value{
numeric or vector
}
\description{
{The well-known density for the hitting time of a given level
in a GBM starting from a known price.}
}
