% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton-hitting-times.R
\name{dhit_merton}
\alias{dhit_merton}
\title{The hitting time PDF of a Merton jump-diffusion}
\usage{
dhit_merton(t, a, param, n = 200)
}
\arguments{
\item{t}{the current time}

\item{a}{the boundary to hit above}

\item{param}{the dynamics of the jump-diffusion, a vector of 5, the drift, volatility
mean-jump rate, mean jump-size, sd of jump size}

\item{n}{the number used in the discretization for differentiating the CDF with
forward differences}
}
\value{
numeric or vector
}
\description{
{The PDF of the hitting time of \code{a} under Merton's jump-diffusion, i.e.
a geometric Levy process with the log being an Ito diffusion plus a compound Poisson
process of Gaussians. Most likely requires smoothing.}
}
