% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{dmvtnorm}
\alias{dmvtnorm}
\title{Multivariate normal probability density function}
\usage{
dmvtnorm(x, mu, Sigma)
}
\arguments{
\item{x}{vector of reals}

\item{mu}{vector means}

\item{Sigma}{covariance matrix}
}
\value{
numeric
}
\description{
{Everybody's favorite distribution, now in multiple dimensions!
(Why isn't this in base R anyway?)}
}
