% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureCoefficients.R
\name{drift_lvm1}
\alias{drift_lvm1}
\title{local drift mixture coefficient function}
\usage{
drift_lvm1(s, t, probs, mus, sigmas, spot)
}
\arguments{
\item{s}{the price}

\item{t}{the time, in trading years}

\item{probs}{the probability of each mixing component}

\item{mus}{the drift of each mixing component}

\item{sigmas}{the volatility of each mixing component}

\item{spot}{the spot value}
}
\value{
numeric
}
\description{
{A local volatility mixture of constant volatilities, with linear behavior near time zero.}
}
\details{
{See .pdf for derivation}
}
