% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable.R
\name{dstable}
\alias{dstable}
\title{Wrapper to libstableR}
\usage{
dstable(x, pars)
}
\arguments{
\item{x}{argument of PDF}

\item{pars}{parameters of stable distribution, see libstableR}
}
\value{
numeric
}
\description{
{A wrapper to \code{libstableR} functions to match
naming conventions of basic pdfs/cdfs/generators.}
}
