% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-fitting.R
\name{ema}
\alias{ema}
\title{Exponentially moving average}
\usage{
ema(x, lambda = 0.94, h = 1/252)
}
\arguments{
\item{x}{data-set}

\item{lambda}{weight parameter (near 0 corresponds to sample mean)}

\item{h}{timescale}
}
\value{
numeric
}
\description{
{Exponentially weighted sample average. Past data has exponentially
decaying weights.
}
}
