% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mertonHittingTimes.R
\name{eta}
\alias{eta}
\title{The exponent used in the method of images for hitting time densities}
\usage{
eta(a, t, param)
}
\arguments{
\item{a}{the boundary to hit above}

\item{t}{the current time}

\item{param}{the dynamics of the jump-diffusion, a vector of 5, the drift, volatility
mean-jump rate, mean jump-size, sd of jump size}
}
\value{
numeric or vector
}
\description{
{The exponent used in the method of images.}
}
