% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-fitting.R
\name{ewmc}
\alias{ewmc}
\title{Exponentially moving covariance}
\usage{
ewmc(X, lambda = 0.94, h = 1/252)
}
\arguments{
\item{X}{data-set}

\item{lambda}{weight parameter (near 0 corresponds to sample mean)}

\item{h}{timescale}
}
\value{
matrix
}
\description{
{Exponentially weighted sample covariance Past data has exponentially
decaying weights.
}
}
