% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{extract_mixture}
\alias{extract_mixture}
\title{Convenience to extract parameters}
\usage{
extract_mixture(mcfit, scale = 252)
}
\arguments{
\item{mcfit}{object returned from \code{Mclust}}

\item{scale}{to scale returns (use only for log-returns data)}
}
\value{
matrix
}
\description{
Convenience to extract parameters
}
