% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-fitting.R
\name{fit_ema_gbm}
\alias{fit_ema_gbm}
\title{Fit GBM model where parameters are estimated using EMAs}
\usage{
fit_ema_gbm(X, lambda = 0.94, h = 1/252)
}
\arguments{
\item{X}{data-set}

\item{lambda}{weight parameter (near 0 corresponds to sample mean)}

\item{h}{timescale}
}
\value{
list of drift and covariance matrix Sigma
}
\description{
{Assuming log-returns. This estimates drift vector and covariance
matrix but with an EMA filter. A log-adjustment is made to the drift per
the GBM dynamics.
}
}
