% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDiscreteTime.R
\name{fitDTFM}
\alias{fitDTFM}
\title{Fit returns distribution to daily-arithmetic returns}
\usage{
fitDTFM(x, distr = "norm")
}
\arguments{
\item{x}{time series of daily arithmetic returns}

\item{distr}{distribution name: "norm" or "unif"}
}
\value{
vector
}
\description{
Fit returns distribution to daily-arithmetic returns
}
