% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{fitGBMs}
\alias{fitGBMs}
\title{Fit correlated geometric Brownian motions to a matrix of daily log-returns time-series}
\usage{
fitGBMs(log_returns, timeScale = 1/252)
}
\arguments{
\item{log_returns}{multi-dimensional time-series of daily log-returns}

\item{timeScale}{time-scale to convert by.}
}
\value{
list containing the vector \code{drift} and matrix \code{Sigma} of point-estimates.
}
\description{
{MLE estimates for the parameters of a vector of correlated GBMs, the mean
drift rate and the volatility coefficient.}
}
