% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heston-fitting.R
\name{fitHeston}
\alias{fitHeston}
\title{fit Heston parameters using particle filter and QMLE}
\usage{
fitHeston(logReturns, iterations, N, nthresh, h)
}
\arguments{
\item{logReturns}{data}

\item{iterations}{number of steps to use in the MLE}

\item{N}{number of particles}

\item{nthresh}{effective threshold for PF}

\item{h}{time-step}
}
\value{
vector
}
\description{
{An ad-hoc Heston fitter.}
}
