% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-family.R
\name{fitMixtureDiffusion}
\alias{fitMixtureDiffusion}
\title{Fit log-normal mixture diffusion to daily-time series of log-returns}
\usage{
fitMixtureDiffusion(log_returns, timescale = 1/252)
}
\arguments{
\item{log_returns}{multi-dimensional time-series of daily log-returns}

\item{timescale}{time-scale to convert by.}
}
\value{
matrix of component parameters, three rows: probabilities, drifts, volatilities
}
\description{
{wrapper to \code{mclust} EM algorithm for the parameters of a Gaussian-mixture diffusion.}
}
