% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton-fitting.R
\name{fitSymmetricMerton}
\alias{fitSymmetricMerton}
\title{Fit the symmetric Merton jump diffusion model to log-return data.}
\usage{
fitSymmetricMerton(x, k = 1/252, numDev = 1)
}
\arguments{
\item{x}{the data-set of daily log returns}

\item{k}{time scale, defaults to 1/252 for daily data}

\item{numDev}{number of daily-standard deviations to use in initial guess for jump sizes.}
}
\value{
output from optim
}
\description{
{Use base R's \code{optim} to maximize the log-likelihood in
the Merton jump diffusion model with symmetric jumps.}
}
