% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heston-fitting.R
\name{hestonLogLikelihood}
\alias{hestonLogLikelihood}
\title{Quasi log-likelihood function for Heston model}
\usage{
hestonLogLikelihood(p, y, v)
}
\arguments{
\item{p}{vector of parameters, see details}

\item{y}{time-series of log-price increments}

\item{v}{time-series of volatility}
}
\value{
numeric
}
\description{
{The log-likelihood function of the Heston model, computed using
Bayesian recursion.}
}
\details{
{The argument \code{p} must be a vector whose entries represent in order,
\eqn{(\kappa, \theta, \xi, \mu)}, the mean-reversion speed, mean-reversion level, the vol-of-vol,
and mean-drift of the stock price.}
}
