% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heston-fitting.R
\name{hestonMLE}
\alias{hestonMLE}
\title{Quasi MLE for Heston dynamics}
\usage{
hestonMLE(y, iterations = 1, N = 100, nthresh = 100)
}
\arguments{
\item{y}{time-series of log-price increments}

\item{iterations}{number of iterations to step through for filtering and}

\item{N}{number of particles to use in the particle-filter}

\item{nthresh}{the threshold of particles in the particle-filter
performing the MLE estimates}
}
\value{
vector/numeric
}
\description{
{An ad-hoc step-wise QMLE routine for estimating parameters
for the Heston dynamics given an observational time-series of log-price increments.}
}
\details{
{At each step, hidden volatility is filtered into an estimate based on log-price observations,
and conditional on these observations, the likelihood is maximized over the parameter space. Then
on the next step, the volatility is re-filtered under the updated parameters until reaching the last observation.
The filtered volatility is tracked step-wise, so that each iteration produces a new filtered-state
appended to the previous filtered-states produced under the previous parameters.}
}
