% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heston-fitting.R
\name{hestonParticleFilter}
\alias{hestonParticleFilter}
\title{Particle filter for filtering volatility in the Heston model}
\usage{
hestonParticleFilter(y, param, N, nthresh, h = 1/252)
}
\arguments{
\item{y}{time-series of log-price increments}

\item{param}{vector of parameters defining Heston dynamics, see details}

\item{N}{number of particles to use}

\item{nthresh}{threshold of particles to use}

\item{h}{the time-step to use}
}
\value{
vector/numeric
}
\description{
{Sequential importance (re)-sampling implementation of the particle method
to estimate the hidden volatility of a Heston model.}
}
\details{
{The argument \code{param} must be a vector whose entries represent in order,
\eqn{(\kappa, \theta, \xi, \mu)}, the mean-reversion speed, mean-reversion level, the vol-of-vol,
and mean-drift of the stock price. The algorithm's details are on Wikipedia.}
}
