% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-hitting-times.R
\name{hitmode_gbm}
\alias{hitmode_gbm}
\title{The mode of the hitting time to a price level for a GBM}
\usage{
hitmode_gbm(gbm, target, spot)
}
\arguments{
\item{gbm}{the parameters defining the GBM, drift and volatility}

\item{target}{the target price}

\item{spot}{the initial price}
}
\value{
numeric
}
\description{
{The mode of the hitting time density for a GBM to
reach a given level.}
}
