% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodRatioTest.R
\name{likelihood_ratio}
\alias{likelihood_ratio}
\title{Likelihood ratio hypothesis test}
\usage{
likelihood_ratio(f, g, alpha1 = 10^-3, alpha2 = 10^-3)
}
\arguments{
\item{f}{denominator density, true under H1}

\item{g}{numerator density, the alternative choice, true under H2}

\item{alpha1}{error of choosing H2 when H1 is true}

\item{alpha2}{error of choosing H1 when H2 is true}
}
\value{
list
}
\description{
{Hypothesis test for choosing among
two densities for fitting data.}
}
