% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeDependentGBM.R
\name{logLikeGBM}
\alias{logLikeGBM}
\title{Log-likelihood function for GBM with time-dependent coefficient functions.}
\usage{
logLikeGBM(theta, x, h, idynamics)
}
\arguments{
\item{theta}{parameter to optimize}

\item{x}{observations of realized variates}

\item{h}{time-scale}

\item{idynamics}{the integrated mean and volatility functions, as a list}
}
\value{
numeric
}
\description{
{The log-likelihood of the daily log returns under a time-dependent
GBM model. Used for numerical MLE routines.}
}
\details{
{Here the coefficients in the SDE \eqn{dS_t=\mu(t) S_t dt +\sigma(t) S_t dB_t} are assumed
to be functions of time.}
}
