% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton.R
\name{logLikeMerton}
\alias{logLikeMerton}
\title{Log-likelihood for Merton jump diffusion log-increments}
\usage{
logLikeMerton(param, x, t)
}
\arguments{
\item{param}{a vector of parameters defining Merton's jump dynamics. See details.}

\item{x}{input value for the CDF}

\item{t}{the time input, for daily use 1/252}
}
\value{
numeric or vector depending on the length of the input \code{x}.
}
\description{
{The log of the PDF of the log-returns in a time interval under Merton's jump diffusion model. Includes
compensator in the drift.}
}
\details{
{The argument \code{param} must be a vector of five real numbers, \eqn{(\mu, \sigma, \lambda, \alpha, \beta)},
representing the mean drift, volatility, mean-rate of jumps, and the mean
log-jump size and the standard deviation of the log-jump size, in this order. The volatility,
mean rate of jumps, and standard deviation of log-jump size must all be positive.}
}
