% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton.R
\name{merton_mle}
\alias{merton_mle}
\title{Estimate parameters for Merton jump-diffusion model via MLE}
\usage{
merton_mle(
  log_returns,
  thresh_hold = 0.02,
  tn = 1,
  time_step = 1/252,
  compensate = FALSE,
  direction = "both"
)
}
\arguments{
\item{log_returns}{log returns of data}

\item{thresh_hold}{threshhold for jump size}

\item{tn}{the length in years of period the data was observed in}

\item{time_step}{time step of data, for real stock data use 1/252, for simulations use \code{tn/n}.}

\item{compensate}{whether to treat the mean as compensated by the jump-drift or not}

\item{direction}{"down", "up" or "both" for direction of jumps to estimate}
}
\value{
list
}
\description{
{Estimate parameters for Merton's jump-diffusion model given a set of variates, using MLE and R's basic \code{optim} function.}
}
