% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-family.R
\name{mixture_vol}
\alias{mixture_vol}
\title{The diffusion coefficient function for a log-normal mixture diffusion}
\usage{
mixture_vol(t, x, param, truncation = 0.5/252)
}
\arguments{
\item{t}{the current time}

\item{x}{the current log-price}

\item{param}{the matrix of component parameters by row, see details.}

\item{truncation}{the truncation of time}
}
\value{
numeric matrix
}
\description{
{The mixture diffusion coefficient is a weighted average of variances.}
}
\details{
{The argument \code{param} must be a square matrix with rows representing,
in order, the probabilities, drifts, and volatilities of each component of the mixture.}
}
