% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton.R
\name{mseMerton}
\alias{mseMerton}
\title{MSE between empirical and model PDF under Merton's jump-diffusion}
\usage{
mseMerton(param, epdf, t)
}
\arguments{
\item{param}{a vector of parameters defining Merton's jump dynamics. See details.}

\item{epdf}{the empirical pdf returned from \code{density}}

\item{t}{the time input, for daily use 1/252}
}
\value{
numeric or vector depending on the length of the input \code{x}.
}
\description{
{MSE between empirical model PDF under Merton's jump-diffusion
for a given time-series of log-returns.}
}
\details{
{The argument \code{param} must be a vector of five real numbers, \eqn{(\mu, \sigma, \lambda, \alpha, \beta)},
representing the mean drift, volatility, mean-rate of jumps, and the mean
log-jump size and the standard deviation of the log-jump size, in this order. The volatility,
mean rate of jumps, and standard deviation of log-jump size must all be positive.}
}
