% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-family.R
\name{pgbm}
\alias{pgbm}
\title{Cumulative distribution function for Black-Scholes distribution}
\usage{
pgbm(s, t, spot, rate, volat)
}
\arguments{
\item{s}{real-numeric input to PDF representing price level}

\item{t}{positive time value}

\item{spot}{the initial spot price}

\item{rate}{the risk-free rate of return}

\item{volat}{the volatility level}
}
\value{
numeric
}
\description{
{The cumulative distribution function for Black-Scholes distributed
variables. }
}
