% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{pgmm}
\alias{pgmm}
\title{Cumulative distribution function for Gaussian mixtures}
\usage{
pgmm(x, p, mus, sigmas)
}
\arguments{
\item{x}{the real variable argument of the PDF}

\item{p}{component probabilities}

\item{mus}{component means}

\item{sigmas}{component standard deviations}
}
\value{
numeric
}
\description{
{A mixture model of PDFs is a probabilistic weighted sum of individual PDFs}
}
