% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton-hitting-times.R
\name{phit_merton}
\alias{phit_merton}
\title{The hitting time CDF of a Merton jump-diffusion}
\usage{
phit_merton(t, a, param)
}
\arguments{
\item{t}{the current time}

\item{a}{the boundary to hit above}

\item{param}{the dynamics of the jump-diffusion, a vector of 5, the drift, volatility
mean-jump rate, mean jump-size, sd of jump size}
}
\value{
numeric or vector
}
\description{
{The CDF of the hitting time of \code{a} under Merton's jump-diffusion, i.e.
a geometric Levy process with the log being an Ito diffusion plus a compound Poisson
process of Gaussians.}
}
