% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kou.R
\name{pkou}
\alias{pkou}
\title{CDF of double-exponential jumps}
\usage{
pkou(x, p, alpha, beta)
}
\arguments{
\item{x}{real-input of the PDF}

\item{p}{mixing probability}

\item{alpha}{mean size of positive jumps}

\item{beta}{mean size of negative jumps}
}
\value{
vector
}
\description{
{CDF of mixture of two exponentials one negative.}
}
\references{
The double-exponential mixture distribution or the Kou distribution was introduced (in mathematical finance) in the paper \href{http://www.columbia.edu/~sk75/MagSci02.pdf}{http://www.columbia.edu/~sk75/MagSci02.pdf} by S.G. Kou.
}
