% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable.R
\name{pstable}
\alias{pstable}
\title{Wrapper to libstableR}
\usage{
pstable(x, pars)

pstable(x, pars)
}
\arguments{
\item{x}{argument of inverse CDF}

\item{pars}{parameters of stable distribution, see libstableR}
}
\value{
numeric

numeric
}
\description{
{A wrapper to \code{libstableR} functions to match
naming conventions of basic pdfs/cdfs/generators.}

{A wrapper to \code{libstableR} functions to match
naming conventions of basic pdfs/cdfs/generators.}
}
