% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable.R
\name{qstable}
\alias{qstable}
\title{Wrapper to libstableR}
\usage{
qstable(p, pars)
}
\arguments{
\item{p}{argument of inverse CDF}

\item{pars}{parameters of stable distribution, see libstableR}
}
\value{
numeric
}
\description{
{A wrapper to \code{libstableR} functions to match
naming conventions of basic pdfs/cdfs/generators.}
}
