% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackscholes.R
\name{rblackscholes}
\alias{rblackscholes}
\title{Generate variates under Black-Scholes distribution}
\usage{
rblackscholes(n, t, spot, rate, volat)
}
\arguments{
\item{n}{number of variates to generate}

\item{t}{positive time value}

\item{spot}{the initial spot price}

\item{rate}{the risk-free rate of return}

\item{volat}{the volatility level}
}
\value{
numeric
}
\description{
{Generate random variates with Black-Scholes distribution.}
}
