% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-family.R
\name{rcornorm}
\alias{rcornorm}
\title{Generate a vector of two correlated standard Gaussian RVs}
\usage{
rcornorm(n, rho)
}
\arguments{
\item{n}{number of vectors to simulate}

\item{rho}{the correlation factor}
}
\value{
numeric vector/matrix
}
\description{
{Uses the Cholesky decomposition on the 2x2 matrix
to simulate a vector of two correlated standard normals.}
}
