% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpois-family.R
\name{rgcpp}
\alias{rgcpp}
\title{Simulate variates under geometric compensated Poisson distributions at a given time.}
\usage{
rgcpp(n, t, spot, a, b, lambda)
}
\arguments{
\item{n}{number of variates to simulate}

\item{t}{time argument to CDF}

\item{spot}{initial spot price}

\item{a}{the coefficient of the Poisson RV}

\item{b}{the coefficient of the compensated drift}

\item{lambda}{the mean rate of arrivals}
}
\value{
numeric
}
\description{
{Simulate variates that follow a geometric compensated Poisson distribution
of a given time and initial spot price. Used for generating terminal values for e.g. option pricing
or solving Feynman-Kac PDEs; not intended for simulating a sample-path.}
}
