% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-family.R
\name{rgmm}
\alias{rgmm}
\title{Sampling algorithm for Gaussian mixtures}
\usage{
rgmm(n, p, mus, sigmas)
}
\arguments{
\item{n}{number of variates to simulate}

\item{p}{the probability vector for component chances}

\item{mus}{the vector of component means}

\item{sigmas}{the vector of volatilities.}
}
\value{
vector
}
\description{
{The arguments can be the component vectors of a mixture model}
}
