% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton-family.R
\name{rmerton}
\alias{rmerton}
\title{Simulate the terminal RV of a log-Merton jump diffusion}
\usage{
rmerton(n, t, param)
}
\arguments{
\item{n}{number of variates to simulate}

\item{t}{the terminal time-horizon}

\item{param}{the parameter set defining the jump diffusion dynamics}
}
\value{
numeric or vector
}
\description{
{Simulates the terminal value of a log jump diffusion
under Merton dynamics.}
}
