% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable.R
\name{stableVAR}
\alias{stableVAR}
\title{Value at Risk for stable distribution}
\usage{
stableVAR(p, pars)
}
\arguments{
\item{p}{level of confidence}

\item{pars}{parameters of stable distribution, see libstableR}
}
\value{
numeric
}
\description{
{Computes the inverse CDF at \code{1-p}.}
}
