% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heston-fitting.R
\name{state_vol}
\alias{state_vol}
\title{Conditional mean and volatility functions of state and observation}
\usage{
state_vol(s, param, h = 1/252)
}
\arguments{
\item{s}{the volatility level}

\item{param}{the vector of parameters, see details}

\item{h}{the time step}
}
\value{
numeric
}
\description{
{The conditional mean and volatility functions for the Heston state and
observation.}
}
\details{
{The argument \code{param} must be a vector whose entries represent in order,
\eqn{(\kappa, \theta, \xi, \mu)}, the mean-reversion speed, mean-reversion level, the vol-of-vol,
and mean-drift of the stock price.}
}
