% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mertonHittingTimes.R
\name{survival_merton}
\alias{survival_merton}
\title{The survival probability of a boundary by a given time under Merton's jump diffusion}
\usage{
survival_merton(t, a, param)
}
\arguments{
\item{t}{the current time}

\item{a}{the boundary to hit above}

\item{param}{the dynamics of the jump-diffusion, a vector of 5, the drift, volatility
mean-jump rate, mean jump-size, sd of jump size}
}
\value{
numeric or vector
}
\description{
{The survival probability of hitting \code{a} by time \code{t}. By survival,
we mean the process is not been stopped yet, i.e. has not hit above the boundary.}
}
