% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton-tests.R
\name{testMertonSimDensity}
\alias{testMertonSimDensity}
\title{Verify the empirical density of Merton simulated variates}
\usage{
testMertonSimDensity(n, t, param)
}
\arguments{
\item{n}{number of variates to simulate}

\item{t}{the terminal time-horizon}

\item{param}{the parameter set defining the jump diffusion dynamics}
}
\value{
vector of errors
}
\description{
{Compares the empirical density of simulated variates against
the exact density per the infinite mixture formula, for Merton log dynamics.}
}
