% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureCoefficients.R
\name{time_homotopy}
\alias{time_homotopy}
\title{Time-homotopy between a constant and a given function of (s,t)}
\usage{
time_homotopy(s, t, f, initial, epsilon = 0.5/252, ...)
}
\arguments{
\item{s}{the current price}

\item{t}{the time}

\item{f}{function to homotopy towards}

\item{initial}{initial constant}

\item{epsilon}{time-cut off}

\item{...}{additional parameters for the function \code{f(s, t, ...)}}
}
\value{
numeric
}
\description{
Time-homotopy between a constant and a given function of (s,t)
}
