% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mertonHittingTimes.R
\name{trans_den_merton}
\alias{trans_den_merton}
\title{The transition density used and reflect in the method of images}
\usage{
trans_den_merton(x, t, a, param)
}
\arguments{
\item{x}{the space variable}

\item{t}{the current time}

\item{a}{the boundary to hit above}

\item{param}{the dynamics of the jump-diffusion, a vector of 5, the drift, volatility
mean-jump rate, mean jump-size, sd of jump size}
}
\value{
numeric or vector
}
\description{
{The transition density figuring in the method of images.}
}
