% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-fitting.R
\name{update_ema}
\alias{update_ema}
\title{Efficient EMA updates}
\usage{
update_ema(x_new, ema_last, alpha, n)
}
\arguments{
\item{x_new}{the new log-return value}

\item{ema_last}{the last ema value}

\item{alpha}{the smoothing factor}

\item{n}{the total size}
}
\value{
vector
}
\description{
Efficient EMA updates
}
