% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-fitting.R
\name{update_ewmc}
\alias{update_ewmc}
\title{Efficient EMA updates}
\usage{
update_ewmc(x_new, Sigma_last, mu_new, alpha, n)
}
\arguments{
\item{x_new}{the new log-return value}

\item{Sigma_last}{the last ema-cov}

\item{mu_new}{the new mean}

\item{alpha}{the smoothing factor}

\item{n}{the total size}
}
\value{
matrix
}
\description{
Efficient EMA updates
}
