% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton.R
\name{dmerton1}
\alias{dmerton1}
\title{The probability density function under Merton's jump-diffusion model for stock-dynamics}
\usage{
dmerton1(x, t, drift, volat, lambda, a, b)
}
\arguments{
\item{x}{Input value for the CDF}

\item{t}{The time input, for daily use 1/252}

\item{drift}{The drift rate}

\item{volat}{the diffusion coefficient}

\item{lambda}{Mean rate of jumps per year}

\item{a}{The mean jump size}

\item{b}{the mean jump volatility}
}
\value{
Numerical
}
\description{
{Auxillary function for PDF for a Jump Diffusion process. Not necessarily under the martingale measure.}
}
