% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-family.R
\name{extract_mixture}
\alias{extract_mixture}
\title{Convenience to extract parameters}
\usage{
extract_mixture(mcfit, timescale = 1/252, continuous = FALSE)
}
\arguments{
\item{mcfit}{object returned from \code{Mclust}}

\item{timescale}{to scale returns (use only for log-returns data)}

\item{continuous}{boolean for discrete time or continuoust time}
}
\value{
matrix
}
\description{
Convenience to extract parameters
}
