% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{fitMixtureDiffusion}
\alias{fitMixtureDiffusion}
\title{Fit correlated geometric Brownian motions to a matrix of daily log-returns time-series}
\usage{
fitMixtureDiffusion(log_returns, timeScale = 1/252)
}
\arguments{
\item{log_returns}{multi-dimensional time-series of daily log-returns}

\item{timeScale}{time-scale to convert by.}
}
\value{
matrix of component parameters, three rows: probabilities, drifts, volatilities
}
\description{
{wrapper to \code{mclust} EM algorithm for the parameters of a Gaussian-mixture diffusion.}
}
