% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable.R
\name{rstable}
\alias{rstable}
\title{Wrapper to libstableR}
\usage{
rstable(n, pars)
}
\arguments{
\item{n}{number of variates to simulate}

\item{pars}{parameters of stable distribution, see libstableR}
}
\value{
numeric
}
\description{
{A wrapper to \code{libstableR} functions to match
naming conventions of basic pdfs/cdfs/generators.}
}
